#189 - Rename executable. Clean up the CMakeLists for testing.
