+++ src/nimble_main.cc
@@ -41,7 +41,7 @@
 //@HEADER
 */
 
-#include "nimble_mpi.h"
+#include "nimble_main.h"
 
 #include "nimble.quanta.stopwatch.h"
 #include "nimble_block_material_interface_factory_base.h"
@@ -257,10 +257,7 @@ NimbleMain(
     if (rve_genesis_file_name != "none") { rve_mesh.ReadFile(rve_genesis_file_name); }
   }
 
-  std::string tag = parser.GetOutputTag();
-#ifdef NIMBLE_HAVE_ARBORX
-  if ((parser.UseKokkos()) && (parser.HasContact())) tag = "arborx";
-#endif
+  std::string tag = "out";
   std::string output_exodus_name = nimble::IOFileName(parser.ExodusFileName(), "e", tag, my_rank, num_ranks);
 
   int dim       = mesh.GetDim();
@@ -390,10 +387,7 @@ ExplicitTimeIntegrator(
     contact_manager->CreateContactEntities(
         mesh, *myVectorCommunicator, contact_primary_block_ids, contact_secondary_block_ids);
     if (contact_visualization) {
-      std::string tag = parser.GetOutputTag();
-#ifdef NIMBLE_HAVE_ARBORX
-      if ((parser.UseKokkos()) && (parser.HasContact())) tag = "arborx";
-#endif
+      std::string tag = "out";
       std::string contact_visualization_exodus_file_name =
           nimble::IOFileName(parser.ContactVisualizationFileName(), "e", tag, my_rank, num_ranks);
       contact_manager->InitializeContactVisualization(contact_visualization_exodus_file_name);
