+++ test/simple_deformation_modes/CMakeLists.txt
@@ -1,35 +1,42 @@
 
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/simple_deformation_modes.in
-               ${CMAKE_CURRENT_BINARY_DIR}/simple_deformation_modes.in COPYONLY)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/simple_deformation_modes.g
-               ${CMAKE_CURRENT_BINARY_DIR}/simple_deformation_modes.g COPYONLY)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/simple_deformation_modes.gold.e
-               ${CMAKE_CURRENT_BINARY_DIR}/simple_deformation_modes.gold.e COPYONLY)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/simple_deformation_modes.exodiff
-               ${CMAKE_CURRENT_BINARY_DIR}/simple_deformation_modes.exodiff COPYONLY)
+set(prefix "simple_deformation_modes")
 
-add_test(simple_deformation_modes_serial python ../run_exodiff_test.py --executable "${NimbleSM_Serial}" --input-deck simple_deformation_modes.in)
+foreach (ext "in" "g" "gold.e" "exodiff")
+  configure_file(${CMAKE_CURRENT_SOURCE_DIR}/${prefix}.${ext}
+                 ${CMAKE_CURRENT_BINARY_DIR}/${prefix}.${ext} COPYONLY)
+endforeach()
 
-if(NIMBLE_HAVE_DARMA)
-  add_test(simple_deformation_modes_darma_np1 python ../run_exodiff_test.py --executable "${NimbleSM_Darma}" --input-deck simple_deformation_modes.in --num-ranks 1 --num-virtual-ranks 1 --charm-run ${charmrun})
+set(inputfile "${prefix}.in")
+
+add_test(NAME "${prefix}-serial"
+         COMMAND python ../run_exodiff_test.py --executable "${nimble_exe}" --cli-flag "" --input-deck "${inputfile}" --num-ranks 1
+        )
+
+if(NIMBLE_HAVE_KOKKOS)
+  add_test(NAME "${prefix}-serial-kokkos"
+           COMMAND python ../run_exodiff_test.py --executable "${nimble_exe}" --cli-flag "use_kokkos" --input-deck "${inputfile}" --num-ranks 1
+          )
 endif()
 
-if(NIMBLE_HAVE_MPI)
-  add_test(simple_deformation_modes_mpi_np1 python ../run_exodiff_test.py --executable "${NimbleSM_MPI}" --input-deck simple_deformation_modes.in --num-ranks 1)
+if (NIMBLE_HAVE_MPI)
+
+  foreach (ext "g.2.0" "g.2.1" "g.4.0" "g.4.1" "g.4.2" "g.4.3")
+    configure_file(${CMAKE_CURRENT_SOURCE_DIR}/${prefix}.${ext}
+                   ${CMAKE_CURRENT_BINARY_DIR}/${prefix}.${ext} COPYONLY)
+  endforeach()
+
+  foreach (nrank 2 4)
+    add_test(NAME "${prefix}-np${nrank}"
+             COMMAND python ../run_exodiff_test.py --executable "${nimble_exe}" --cli-flag "" --input-deck "${inputfile}" --num-ranks ${nrank}
+            )
+
+    if(NIMBLE_HAVE_KOKKOS)
+      add_test(NAME "${prefix}-np${nrank}-kokkos"
+               COMMAND python ../run_exodiff_test.py --executable "${nimble_exe}" --cli-flag "use_kokkos" --input-deck "${inputfile}" --num-ranks ${nrank}
+              )
+    endif()
+
+  endforeach()
+
 endif()
 
-if(NIMBLE_HAVE_ARBORX)
-  add_test(simple_deformation_modes_arborx_np1 python ../run_exodiff_test.py --executable "${NimbleSM_ArborX}" --input-deck simple_deformation_modes.in --num-ranks 1)
-endif()
-
-if(NIMBLE_HAVE_KOKKOS AND NOT NIMBLE_HAVE_ARBORX)
-  add_test(simple_deformation_modes_kokkos_np1 python ../run_exodiff_test.py --executable "${NimbleSM_Kokkos}" --input-deck simple_deformation_modes.in --num-ranks 1)
-endif()
-
-if(NIMBLE_HAVE_TRILINOS)
-  add_test(simple_deformation_modes_tpetra_np1 python ../run_exodiff_test.py --executable "${NimbleSM_Tpetra}" --input-deck simple_deformation_modes.in --num-ranks 1)
-endif()
-
-if(NIMBLE_HAVE_QTHREADS)
-  add_test(simple_deformation_modes_qthreads_np1_ns1_nwps1 python ../run_exodiff_test.py --executable "${NimbleSM_Qthreads}" --input-deck simple_deformation_modes.in --num-ranks 1 --qthreads-num-shepherds 1 --qthreads-num-workers-per-shepherd 1)
-endif()
